Cleanup: move extension to separate file.
